+++ agent/temboardagent/cli/runtask.py
@@ -1,8 +1,16 @@
-import logging
+# Note that our home-made background task implementation, we use a different
+# semantic that state of the art background task implemtation:
+#
+# a task function is called a worker
+# a worker process is called workerpool
+# a message is called a task
+# the message broker is called taskmanager
+#
+
+import logging.config
 
 from ..toolkit.app import SubCommand
 from ..toolkit.taskmanager import RunTaskMixin
-from ..model import check_schema
 from .app import app
 
 
@@ -20,6 +28,5 @@ class RunTask(RunTaskMixin, SubCommand):
             self.print_workers(workers)
         else:
             worker, worker_args = self.compute_worker_args(workers, args)
-            check_schema()
             worker(*worker_args)
         return 0