+++ ui/temboardui/cli/app.py
@@ -34,6 +34,9 @@ logger = logging.getLogger('temboardui')
 
 
 class TemboardApplication(BaseApplication):
+    PROGRAM = 'temboard'
+    VERSION = __version__
+
     DEFAULT_CONFIGFILES = [
         '/etc/temboard/temboard.conf',
         'temboard.conf',
@@ -46,9 +49,6 @@ class TemboardApplication(BaseApplication):
         'maintenance',
         'statements',
     ]
-    PROGRAM = 'temboard'
-    REPORT_URL = "https://github.com/dalibo/temboard/issues"
-    VERSION = __version__
 
     def define_arguments(self, parser):
         define_core_arguments(parser)
@@ -121,7 +121,7 @@ class TemboardApplication(BaseApplication):
 
         command_name = getattr(args, 'command_fullname', 'serve')
         command = self.commands[command_name]
-        command.main(args)
+        return command.main(args)
 
     def apply_config(self):
         bootstrap_tornado = not hasattr(self, 'webapp')